+++ dart/lib/src/http_client/tracing_client.dart
@@ -46,15 +46,28 @@ class TracingClient extends BaseClient {
 
     StreamedResponse? response;
     try {
-      if (span != null) {
-        if (containsTargetOrMatchesRegExp(
-            _hub.options.tracePropagationTargets, request.url.toString())) {
-          addSentryTraceHeader(span, request.headers);
-          addBaggageHeader(
+      if (containsTargetOrMatchesRegExp(
+          _hub.options.tracePropagationTargets, request.url.toString())) {
+        if (span != null) {
+          addSentryTraceHeaderFromSpan(span, request.headers);
+          addBaggageHeaderFromSpan(
             span,
             request.headers,
             logger: _hub.options.logger,
           );
+        } else {
+          final scope = _hub.scope;
+          final propagationContext = scope.propagationContext;
+
+          final traceHeader = propagationContext.toSentryTrace();
+          addSentryTraceHeader(traceHeader, request.headers);
+
+          final baggage = propagationContext.baggage;
+          if (baggage != null) {
+            final baggageHeader = SentryBaggageHeader.fromBaggage(baggage);
+            addBaggageHeader(baggageHeader, request.headers,
+                logger: _hub.options.logger);
+          }
         }
       }