+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add db.system and db.name attributes to db spans data ([#1629](https://github.com/getsentry/sentry-dart/pull/1629))
 
+### Features
+
+- Tracing without performance ([#1621](https://github.com/getsentry/sentry-dart/pull/1621))
+
 ### Fixes
 
 - Normalize data properties of `SentryUser` and `Breadcrumb` before sending over method channel ([#1591](https://github.com/getsentry/sentry-dart/pull/1591))